# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Parkings",
        "description": "API REST desarrollada en Laravel para la gestión de parkings.",
        "contact": {
            "email": "tu-email@ejemplo.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pedro Geraghty"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pedro@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión y obtener token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pedro@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|wS7RzG0y..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/parkings": {
            "post": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Crear un nuevo parking",
                "operationId": "2345ce2d01d1cd69e0c3eef1efcd0a1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Parking Centro"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Av. Corrientes 123"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -34.6037
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -58.3816
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/parkings/{id}": {
            "get": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Obtener un parking por ID",
                "operationId": "b6ad1cab1d4056a1261f289c7d0caeea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del parking a consultar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parking encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Parking Centro"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Av. Corrientes 123"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -34.6037
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": -58.3816
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parking no encontrado"
                    }
                }
            }
        },
        "/api/parkings/closest": {
            "get": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Obtener el parking más cercano a una ubicación",
                "operationId": "6a8e23b19ee6645d25fde561abb2542f",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitud del punto de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -34.6037
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitud del punto de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -58.3816
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parking más cercano",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "distance_m": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 320.5
                                        },
                                        "parking": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Parking Centro"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Av. Corrientes 123"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -34.6037
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -58.3816
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron parkings"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Parkings",
            "description": "Parkings"
        }
    ]
}